snmp_view_based_acm_mib (snmp v5.15)

Instrumentation Functions for SNMP-VIEW-BASED-ACM-MIB

The module snmp_view_based_acm_mib implements the instrumentation functions for the SNMP-VIEW-BASED-ACM-MIB, and functions for configuring the database.

The configuration files are described in the SNMP User's Manual.

Types

external_view_mask/0

-type external_view_mask() :: octet_string().

internal_view_mask/0

-type internal_view_mask() :: null | [internal_view_mask_element()].

internal_view_mask_element/0

-type internal_view_mask_element() :: 0 | 1.

internal_view_type/0

-type internal_view_type() :: 1 | 2.

mibview/0

-type mibview() :: [{SubTree :: snmp:oid(), Mask :: internal_view_mask(), Type :: internal_view_type()}].

octet/0

-type octet() :: byte().

octet_string/0

-type octet_string() :: [octet()].

Functions

add_access(GroupName, Prefix, SecModel, SecLevel, Match, RV, WV, NV)

add_access(GroupName, Prefix, SecModel, SecLevel, Match, RV, WV, NV) -> Ret

Adds a access definition to the agent config. Equivalent to one vacmAccess-line in the vacm.conf file.

add_sec2group(SecModel, SecName, GroupName)

add_sec2group(SecModel, SecName, GroupName) -> Ret

Adds a security to group definition to the agent config. Equivalent to one vacmSecurityToGroup-line in the vacm.conf file.

add_view_tree_fam(ViewIndex, SubTree, Status, Mask)

add_view_tree_fam(ViewIndex, SubTree, Status, Mask) -> Ret

Adds a view tree family definition to the agent config. Equivalent to one vacmViewTreeFamily-line in the vacm.conf file.

configure(Dir)

configure(ConfDir) -> void()

This function is called from the supervisor at system start-up.

Inserts all data in the configuration files into the database and destroys all old rows with StorageType volatile. The rows created from the configuration file will have StorageType nonVolatile.

All snmp counters are set to zero.

If an error is found in the configuration file, it is reported using the function config_err/2 of the error report module, and the function fails with the reason configuration_error.

ConfDir is a string which points to the directory where the configuration files are found.

The configuration file read is: vacm.conf.

delete_access(Key)

delete_access(Key) -> Ret

Delete a access definition from the agent config.

delete_sec2group(Key)

delete_sec2group(Key) -> Ret

Delete a security to group definition from the agent config.

delete_view_tree_fam(Key)

delete_view_tree_fam(Key) -> Ret

Delete a view tree family definition from the agent config.

reconfigure(Dir)

reconfigure(ConfDir) -> void()

Inserts all data in the configuration files into the database and destroys all old data, including the rows with StorageType nonVolatile. The rows created from the configuration file will have StorageType nonVolatile.

Thus, the data in the SNMP-VIEW-BASED-ACM-MIB, after this function has been called, is the data from the configuration files.

All snmp counters are set to zero.

If an error is found in the configuration file, it is reported using the function config_err/2 of the error report module, and the function fails with the reason configuration_error.

ConfDir is a string which points to the directory where the configuration files are found.

The configuration file read is: vacm.conf.