snmpa_error_logger (snmp v5.19.1)

Functions for Reporting SNMP Errors through the error_logger

The module snmpa_error_logger implements the snmpa_error_report behaviour (see snmpa_error_report) containing two callback functions which are called in order to report SNMP errors.

This module provides a simple mechanism for reporting SNMP errors. Errors are sent to the error_logger after a size check. Messages are truncated after 1024 chars. It is provided as an example.

This module is the default error report module, but can be explicitly configured, see snmpa_error and configuration parameters.

See Also

error_logger(3)

Functions

config_err(Format, Args)

-spec config_err(Format, Args) -> snmp:void() when Format :: string(), Args :: list().

The function is called if an error occurs during the configuration phase, for example if a syntax error is found in a configuration file.

Format and Args are as in io:format(Format, Args).

user_err(Format, Args)

-spec user_err(Format, Args) -> snmp:void() when Format :: string(), Args :: list().

The function is called if a user related error occurs at run-time, for example if a user defined instrumentation function returns erroneous.

Format and Args are as in io:format(Format, Args).