snmpa_mpd (snmp v5.15)

Message Processing and Dispatch module for the SNMP agent

The module snmpa_mpd implements the version independent Message Processing and Dispatch functionality in SNMP for the agent. It is supposed to be used from a Network Interface process (Definition of Agent Net if).

DATA TYPES

See the data types in snmpa_conf.

Functions

discarded_pdu/1

discarded_pdu(Variable) -> void()

Increments the variable associated with a discarded pdu. This function can be used when the net_if process receives a discarded_pdu message from the agent.

generate_msg/5
(since OTP R14B)

Equivalent to generate_msg/6

generate_msg/6
(since OTP R14B)

generate_msg(Vsn, NoteStore, Pdu, MsgData, LocalEngineID, To) -> {ok, PacketsAndAddresses} | {discarded, Reason}

Generates a possibly encrypted request packet to be sent to the network.

MsgData is the message specific data used in the SNMP message. This value is received in a send_pdu or send_pdu_req message from the agent. In SNMPv1 and SNMPv2c, this message data is the community string. In SNMPv3, it is the context information.

To is a list of destination addresses and their corresponding security parameters. This value is received in the same message from the agent and then transformed through process_taddrs before passed to this function.

Note

Note that the use of the LocalEngineID argument is only intended for special cases, if the agent is to "emulate" multiple EngineIDs! By default, the agent uses the value of SnmpEngineID (see SNMP-FRAMEWORK-MIB).

generate_response_msg/5
(since OTP R14B)

Equivalent to generate_response_msg/6

generate_response_msg(Vsn, RePdu, Type, ACMData, Log, N)
(since OTP R14B)

generate_response_msg(Vsn, RePdu, Type, ACMData, LocalEngineID, Log) -> {ok, Packet} | {discarded, Reason}

Generates a possibly encrypted response packet to be sent to the network. Type is the #pdu.type of the original request.

Note

Note that the use of the LocalEngineID argument is only intended for special cases, if the agent is to "emulate" multiple EngineIDs! By default, the agent uses the value of SnmpEngineID (see SNMP-FRAMEWORK-MIB).

init(Vsns)

init(Vsns) -> mpd_state()

This function can be called from the net_if process at start-up. The options list defines which versions to use.

It also initializes some SNMP counters.

process_packet/5
(since OTP 17.3,OTP R14B)

Equivalent to process_packet/6

process_packet(Packet, Domain, Address, State, NoteStore, Log)
(since OTP 17.3,OTP R14B)

process_packet(Packet, From, LocalEngineID, State, NoteStore, Log) -> {ok, Vsn, Pdu, PduMS, ACMData} | {discarded, Reason} | {discovery, DiscoPacket}

Processes an incoming packet. Performs authentication and decryption as necessary. The return values should be passed to the agent.

Note

Note that the use of the LocalEngineID argument is only intended for special cases, if the agent is to "emulate" multiple EngineIDs! By default, the agent uses the value of SnmpEngineID (see SNMP-FRAMEWORK-MIB).

process_taddrs(Dests)
(since OTP 17.3)

process_taddrs(TDests) -> Dests

Transforms addresses from internal MIB format to one more useful to Agent Net if.

See also generate_msg.