snmpa_notification_filter behaviour (snmp v5.15)

Behaviour module for the SNMP agent notification filters.

This module defines the behaviour of the agent notification filters. A snmpa_notification_filter compliant module must export the following functions:

handle_notification/2

The semantics of them and their exact signatures are explained below.

The purpose of notification filters is to allow for modification and/or suppression of a notification.

A misbehaving filter will be removed.

Types

notification/0

-type notification() :: term().

trap/0

-type trap() :: term().

Callbacks

handle_notification(Notif, Data)

-callback handle_notification(Notif, Data) -> Reply
                       when
                           Reply :: send | {send, NewNotif} | dont_send,
                           Notif :: notification() | trap(),
                           NewNotif :: notification() | trap(),
                           Data :: term().

Handle a notification to be sent. The filter can either accept the notification as is, return send, modify the notification, return {send, NewNotif} or suppress the notification, return dont_send.

Data is supplied at filter registration time, see register_notification_filter.