snmp_community_mib (snmp v5.9.1)

The module snmp_community_mib implements the instrumentation functions for the SNMP-COMMUNITY-MIB, and functions for configuring the database.

The configuration files are described in the SNMP User's Manual.

Functions

add_community(Idx, CommName, SecName, CtxName, TransportTag) -> Ret
(since OTP R14B03)

Idx = string()

CommName = string()

SecName = string()

EngineId = string()

CtxName = string()

TransportTag = string()

Ret = {ok, Key} | {error, Reason}

Key = term()

Reason = term()

Adds a community to the agent config. Equivalent to one line in the community.conf file.

With the EngineId argument it is possible to override the configured engine-id (SNMP-FRAMEWORK-MIB).

add_community(Idx, CommName, SecName, EngineId, CtxName, TransportTag) -> Ret
(since OTP R14B03)

configure(ConfDir) -> void()

ConfDir = string()

This function is called from the supervisor at system start-up.

Inserts all data in the configuration files into the database and destroys all old rows with StorageType volatile. The rows created from the configuration file will have StorageType nonVolatile.

All snmp counters are set to zero.

If an error is found in the configuration file, it is reported using the function config_err/2 of the error, report module and the function fails with reason configuration_error.

ConfDir is a string which points to the directory where the configuration files are found.

The configuration file read is: community.conf.

delete_community(Key) -> Ret

Key = term()

Ret = ok | {error, Reason}

Reason = term()

Delete a community from the agent config.

reconfigure(ConfDir) -> void()

ConfDir = string()

Inserts all data in the configuration files into the database and destroys all old data, including the rows with StorageType nonVolatile. The rows created from the configuration file will have StorageType nonVolatile.

Thus, the data in the SNMP-COMMUNITY-MIB, after this function has been called, is from the configuration files.

All snmp counters are set to zero.

If an error is found in the configuration file, it is reported using the function config_err/2 of the error report module, and the function fails with reason configuration_error.

ConfDir is a string which points to the directory where the configuration files are found.

The configuration file read is: community.conf.