snmp_framework_mib (snmp v5.15)

Instrumentation Functions for SNMP-FRAMEWORK-MIB

The module snmp_framework_mib implements instrumentation functions for the SNMP-FRAMEWORK-MIB, and functions for initializing and configuring the database.

The configuration files are described in the SNMP User's Manual.

Types

admin_string()

-type admin_string() :: string().

OCTET STRING (SIZE(0..255))

engine_id()

-type engine_id() :: string().

OCTET STRING (SIZE(5..32))

Functions

add_context(Ctx)

-spec add_context(Ctx) -> {ok, Key} | {error, Reason}
               when Ctx :: string(), Key :: term(), Reason :: term().

Adds a context to the agent config. Equivalent to one line in the context.conf file.

configure(ConfDir)

-spec configure(ConfDir) -> snmp:void() when ConfDir :: string().

This function is called from the supervisor at system start-up.

Inserts all data in the configuration files into the database and destroys all old data.

Thus, the data in the SNMP-FRAMEWORK-MIB, after this function has been called, is from the configuration files.

All snmp counters are set to zero.

If an error is found in the configuration file, it is reported using the function config_err/2 of the error report module, and the function fails with reason configuration_error.

ConfDir is a string which points to the directory where the configuration files are found.

The configuration file read is: context.conf.

delete_context(Key)

-spec delete_context(Key) -> ok | {error, Reason} when Key :: term(), Reason :: term().

Delete a context from the agent config.

init()

-spec init() -> snmp:void().

This function is called from the supervisor at system start-up.

Creates the necessary objects in the database if they do not exist. It does not destroy any old values.