snmp_target_mib (snmp v5.9.1)

The module snmp_target_mib implements the instrumentation functions for the SNMP-TARGET-MIB, and functions for configuring the database.

The configuration files are described in the SNMP User's Manual.

Legacy API functions add_addr/10 that does not specify transport domain, and add_addr/11 that has got separate IpAddr and PortNumber arguments still work as before for backwards compatibility reasons.

DATA TYPES

See the data types in snmpa_conf.

Functions

add_addr(Name, Domain, Addr, Timeout, Retry, TagList, Params, EngineId, TMask, MMS) -> Ret

Name = string()

Domain = transportDomain()

Addr = transportAddress() % Default port is 162

Timeout = integer()

Retry = integer()

TagList = string()

ParamsName = string()

EngineId = string()

TMask = transportAddressMask() % Depends on Domain

MMS = integer()

Ret = {ok, Key} | {error, Reason}

Key = term()

Reason = term()

Adds a target address definition to the agent config. Equivalent to one line in the target_addr.conf file.

add_params(Name, MPModel, SecModel, SecName, SecLevel) -> Ret

Name = string()

MPModel = v1 | v2c | v3

SecModel = v1 | v2c | usm

SecName = string()

SecLevel = noAuthNoPriv | authNoPriv | authPriv

Ret = {ok, Key} | {error, Reason}

Key = term()

Reason = term()

Adds a target parameter definition to the agent config. Equivalent to one line in the target_params.conf file.

configure(ConfDir) -> void()

ConfDir = string()

This function is called from the supervisor at system start-up.

Inserts all data in the configuration files into the database and destroys all old rows with StorageType volatile. The rows created from the configuration file will have StorageType nonVolatile.

All snmp counters are set to zero.

If an error is found in the configuration file, it is reported using the function config_err/2 of the error report module, and the function fails with the reason configuration_error.

ConfDir is a string which points to the directory where the configuration files are found.

The configuration files read are: target_addr.conf and target_params.conf.

delete_addr(Key) -> Ret

Key = term()

Ret = ok | {error, Reason}

Reason = term()

Delete a target address definition from the agent config.

delete_params(Key) -> Ret

Key = term()

Ret = ok | {error, Reason}

Reason = term()

Delete a target parameter definition from the agent config.

reconfigure(ConfDir) -> void()

ConfDir = string()

Inserts all data in the configuration files into the database and destroys all old data, including the rows with StorageType nonVolatile. The rows created from the configuration file will have StorageType nonVolatile.

Thus, the data in the SNMP-TARGET-MIB, after this function has been called, is the data from the configuration files.

All snmp counters are set to zero.

If an error is found in the configuration file, it is reported using the function config_err/2 of the , and the function fails with the reason configuration_error.

ConfDir is a string which points to the directory where the configuration files are found.

The configuration files read are: target_addr.conf and target_params.conf.

set_target_engine_id(TargetAddrName, EngineId) -> boolean()

TargetAddrName = string()

EngineId = string()

Changes the engine id for a target in the snmpTargetAddrTable. If notifications are sent as Inform requests to a target, its engine id must be set.