snmpm_mpd (snmp v5.9.1)

The module snmpm_mpd implements the version independent Message Processing and Dispatch functionality in SNMP for the manager. It is supposed to be used from a Network Interface process (Definition of Manager Net if).

Legacy API function process_msg/7 that has got separate IpAddr and PortNumber arguments still works as before for backwards compatibility reasons.

Functions

generate_msg(Vsn, NoteStore, Pdu, MsgData, Logger) -> {ok, Packet} | {discarded, Reason}

Vsn = 'version-1' | 'version-2' | 'version-3'

NoteStore = pid()

Pdu = #pdu

MsgData = term()

Logger = function()

Packet = binary()

Reason = term()

Generates a possibly encrypted packet to be sent to the network.

NoteStore is the pid() of the note-store process.

MsgData is the message specific data used in the SNMP message. In SNMPv1 and SNMPv2c, this message data is the community string. In SNMPv3, it is the context information.

Logger is the function used for audit trail logging.

generate_response_msg(Vsn, Pdu, MsgData, Logger) -> {ok, Packet} | {discarded, Reason}

Vsn = 'version-1' | 'version-2' | 'version-3'

Pdu = #pdu

MsgData = term()

Logger = function()

Packet = binary()

Reason = term()

Generates a possibly encrypted response packet to be sent to the network.

MsgData is the message specific data used in the SNMP message. This value is received from the process_msg function.

init(Vsns) -> mpd_state()

Vsns = [Vsn]

Vsn = v1 | v2 | v3

This function can be called from the net_if process at start-up. The options list defines which versions to use.

It also initializes some SNMP counters.

process_msg(Msg, Domain, Addr, State, NoteStore, Logger) -> {ok, Vsn, Pdu, PduMS, MsgData} | {discarded, Reason}
(since OTP 17.3)

Msg = binary()

Domain = transportDomainUdpIpv4 | transportDomainUdpIpv6

Addr = {inet:ip_address(), inet:port_number()}

State = mpd_state()

NoteStore = pid()

Logger = function()

Vsn = 'version-1' | 'version-2' | 'version-3'

Pdu = #pdu

PduMs = integer()

MsgData = term()

Processes an incoming message. Performs authentication and decryption as necessary. The return values should be passed the manager server.

NoteStore is the pid() of the note-store process.

Logger is the function used for audit trail logging.

In the case when the pdu type is report, MsgData is either ok or {error, ReqId, Reason}.