snmpm_network_interface behaviour (snmp v5.16)

Behaviour module for the SNMP manager network interface.

This module defines the behaviour of the manager network interface. A snmpm_network_interface compliant module must export the following functions:

start_link/2

stop/1

send_pdu/7

inform_response/4

note_store/2

info/1

get_log_type/1

set_log_type/2

verbosity/2

The semantics of them and their exact signatures are explained below.

Legacy API function send_pdu/7 that has got separate IpAddr and PortNumber arguments still works as before for backwards compatibility reasons.

Callbacks

get_log_type(Pid)

-callback get_log_type(Pid) -> {ok, LogType} | {error, Reason}
                when Pid :: pid(), LogType :: snmp:atl_type(), Reason :: term().

The Audit Trail Log is managed by the network interface process. So, it is this process that has to return the actual log-type.

info(Pid)

-callback info(Pid) -> Info when Pid :: pid(), Info :: [{Key, Value}], Key :: term(), Value :: term().

The info returned is basically up to the implementer to decide. The implementation provided by this application provides info about memory allocation and various socket information.

The info returned by this function is returned together with other info collected by the manager when the info function is called (tagged with the key net_if).

inform_response(Pid, Ref, Addr, Port)

-callback inform_response(Pid, Ref, Addr, Port) -> snmp:void()
                   when
                       Pid :: pid(),
                       Ref :: term(),
                       Addr :: inet:ip_address(),
                       Port :: inet:port_number().

Instruct the network interface process to send the response (acknowledgment) to an inform-request.

Ref is something that can be used to identify the inform-request, e.g. request-id of the inform-request.

Addr and Port identifies the agent, from which the inform-request originated.

note_store(Pid, NoteStore)

-callback note_store(Pid, NoteStore) -> snmp:void() when Pid :: pid(), NoteStore :: pid().

Change the pid of the note-store process. This is used when the server re-starts the note_store (e.g. after a crach).

send_pdu(Pid, Pdu, Vsn, MsgData, Domain, Addr, ExtraInfo)

-callback send_pdu(Pid, Pdu, Vsn, MsgData, Domain, Addr, ExtraInfo) -> snmp:void()
            when
                Pid :: pid(),
                Pdu :: snmp:pdu(),
                Vsn :: 'version-1' | 'version-2' | 'version-3',
                MsgData :: term(),
                Domain :: snmp:tdomain(),
                Addr :: {inet:ip_address(), inet:port_number()},
                ExtraInfo :: term().

Request the network interface process (Pid) to send this pdu (Pdu).

ExtraInfo is some opaque data that is passed to the net-if process. It originates from the ExtraInfo parameter in the calls to the synchronous get-request, asynchronous get-request, synchronous get-next-request, asynchronous get-next-request, synchronous set-request and asynchronous set-request functions. Whether the net-if process chooses to use this is implementation dependent. The net-if process included in this application ignores it.

set_log_type(Pid, NewType)

-callback set_log_type(Pid, NewType) -> {ok, OldType} | {error, Reason}
                when
                    Pid :: pid(),
                    NewType :: snmp:atl_type(),
                    OldType :: snmp:atl_type(),
                    Reason :: term().

The Audit Trail Log is managed by the network interface process. So, it is this process that has to do the actual changing of the type.

See snmpm:set_log_type/1 for more info.

start_link(Server, NoteStore)

-callback start_link(Server, NoteStore) -> {ok, Pid} | {error, Reason}
              when Server :: pid(), NoteStore :: pid(), Pid :: pid(), Reason :: term().

Start-link the network interface process.

Server is the pid of the managing process.

NoteStore is the pid of the note-store process.

stop(Pid)

-callback stop(Pid) -> snmp:void() when Pid :: pid().

Stop the network interface process.

verbosity(Pid, Verbosity)

-callback verbosity(Pid, Verbosity) -> snmp:void() when Pid :: pid(), Verbosity :: snmp:verbosity().

Change the verbosity of the network interface process.